backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.1.2
cryptography>=41.0.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS and middleware
python-dotenv>=1.0.0

# Reuse CLI dependencies
nba-api>=1.10.0
rich==13.7.0
requests-oauthlib==1.3.1
prompt-toolkit==3.0.43

# Yahoo Fantasy Sports
yfpy

# Timezone handling
pytz>=2024.1

# Development dependencies (linting and testing)
pylint==3.0.3
mypy==1.8.0
pytest>=7.4.0

